feat: Add treatments functions
