+++ express/src/server.ts
@@ -1,5 +1,3 @@
-/** @format */
-
 import express, { Express } from "express";
 import { Server } from "socket.io";
 import http from "http";
@@ -8,9 +6,24 @@ import pkg from "pg";
 import dotenv from "dotenv";
 import bcrypt from "bcryptjs";
 import * as fs from "fs";
-import { CONNECTED_USERS, CONNECTION, CREATE_RECORD, DELETE_RECORD, JOIN_ROOM, LOCK_RECORD, SAVE_RECORD, USER_JOINED, UNLOCK_RECORD, SCHEDULING_ROOM_ID, PORT, ARCHIVE_ROOM_ID } from "shared";
+import {
+  CREATE_RECORD,
+  DELETE_RECORD,
+  JOIN_ROOM,
+  LOCK_RECORD,
+  SAVE_RECORD,
+  USER_JOINED,
+  UNLOCK_RECORD,
+  SCHEDULING_ROOM_ID,
+  PORT,
+  ARCHIVE_ROOM_ID,
+  KEY_OF_SERVER_TREATMENT,
+  KEY_OF_SERVER_PRECORD,
+  CONNECTED_USERS,
+  CONNECTION,
+} from "shared";
 import { deconstructRecord, lockOrUnlockRowsQuery, setUserSessionQuery, deconstructTreatement, updateQuery } from "./utils.js";
-import { KEY_OF_SERVER_PRECORD, KEY_OF_SERVER_TREATMENT, TREATMENTS } from "./contants.js";
+import { TREATMENTS } from "./contants.js";
 
 dotenv.config();
 
@@ -348,7 +361,6 @@ app.delete("/api/deleteTreatment", async (req, res) => {
 
   try {
     const q = `DELETE FROM ${tag}.${TREATMENTS} WHERE tr_id=${id}`;
-    console.log(q);
     const result = await pool.query(q);
     res.status(200).json(result);
   } catch (error: any) {
@@ -365,19 +377,18 @@ app.post("/api/insertTreatment", async (req, res) => {
   }
 
   try {
-    let fieldnames = "(";
-    let values = "(";
-    for (let i = 1; i < KEY_OF_SERVER_TREATMENT.length; i++) {
-      fieldnames += `${KEY_OF_SERVER_TREATMENT[i]},`;
-      values += `$${i}`;
-    }
-    values += ")";
-    fieldnames += ")";
+    await pool.connect();
 
-    const q = `INSERT INTO ${tag}.${TREATMENTS} ${fieldnames} VALUES ${values} RETURNING *`;
-    const result = await pool.query(q);
-
-    res.status(200).json(result);
+    // Step 1: Get the current maximum tr_id from the table
+    const maxIdResult = await pool.query("SELECT MAX(tr_id) AS max_id FROM gn_ss_bailor.TREATMENTS");
+    let maxId = maxIdResult.rows[0].max_id || 0;
+    maxId += 1;
+    // Step 3: Insert the new row with DEFAULT VALUES and return the result
+    const insertQuery = `
+      INSERT INTO gn_ss_bailor.TREATMENTS (tr_id) VALUES (${maxId}) RETURNING *;
+    `;
+    const insertResult = await pool.query(insertQuery);
+    res.status(200).json(insertResult);
   } catch (error: any) {
     console.log(error);
   }
