+++ express/src/utils.ts
@@ -1,6 +1,6 @@
 /** @format */
 
-import { KEY_OF_CLIENT_PRECORD, KEY_OF_CLIENT_TREATMENT } from "./contants.js";
+import { KEY_OF_CLIENT_PRECORD, KEY_OF_CLIENT_TREATMENT } from "shared";
 
 export const deconstructRecord = (record: any) => {
   const newRecord: any = {};
@@ -45,7 +45,6 @@ export const deconstructTreatement = (treatment: any) => {
   for (const s of KEY_OF_CLIENT_TREATMENT) {
     newTreatment[s] = treatment[s];
   }
-  console.log(newTreatment);
 
   const values = Object.values(newTreatment);
   if (values && values.length > 0) {
